pyproject.toml
[project]
name = "dinowm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "einops>=0.8.1",
    "timm>=1.0.19",
    "torch>=2.8.0",
    "torchvision>=0.18.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "matplotlib>=3.7.0",
    "gym>=0.26.0",
    "scipy>=1.11.0",
    "opencv-python>=4.8.0",
]
